Extract enterprise runtime configuration module
