+++ app/enterprise_readiness.py
@@ -1,11 +1,50 @@
-import json
 import logging
-import os
 from datetime import datetime, timezone
 from typing import Any, Awaitable, Callable, Mapping
 
 from fastapi import Request, Response
 from fastapi.responses import JSONResponse
+
+from app import enterprise_runtime_config as _runtime_config
+
+_DEFAULT_ENTERPRISE_POLICY_VERSION = _runtime_config._DEFAULT_ENTERPRISE_POLICY_VERSION
+_DEFAULT_MAX_WRITE_PAYLOAD_BYTES = _runtime_config._DEFAULT_MAX_WRITE_PAYLOAD_BYTES
+_DIAGNOSTIC_LIST_SEPARATOR = _runtime_config._DIAGNOSTIC_LIST_SEPARATOR
+_EMPTY_ENV_VALUE = _runtime_config._EMPTY_ENV_VALUE
+_EMPTY_JSON_OBJECT = _runtime_config._EMPTY_JSON_OBJECT
+_ENV_ENABLED_VALUES = _runtime_config._ENV_ENABLED_VALUES
+_ENV_ENTERPRISE_CAPABILITY_RULES_JSON = _runtime_config._ENV_ENTERPRISE_CAPABILITY_RULES_JSON
+_ENV_ENTERPRISE_ENFORCE_AUTHZ = _runtime_config._ENV_ENTERPRISE_ENFORCE_AUTHZ
+_ENV_ENTERPRISE_ENFORCE_PRIVILEGED_READ_AUTHZ = _runtime_config._ENV_ENTERPRISE_ENFORCE_PRIVILEGED_READ_AUTHZ
+_ENV_ENTERPRISE_ENFORCE_RUNTIME_CONFIG = _runtime_config._ENV_ENTERPRISE_ENFORCE_RUNTIME_CONFIG
+_ENV_ENTERPRISE_FEATURE_FLAGS_JSON = _runtime_config._ENV_ENTERPRISE_FEATURE_FLAGS_JSON
+_ENV_ENTERPRISE_MAX_WRITE_PAYLOAD_BYTES = _runtime_config._ENV_ENTERPRISE_MAX_WRITE_PAYLOAD_BYTES
+_ENV_ENTERPRISE_POLICY_VERSION = _runtime_config._ENV_ENTERPRISE_POLICY_VERSION
+_ENV_ENTERPRISE_PRIMARY_KEY_ID = _runtime_config._ENV_ENTERPRISE_PRIMARY_KEY_ID
+_ENV_ENTERPRISE_PRIVILEGED_READ_RULES_JSON = _runtime_config._ENV_ENTERPRISE_PRIVILEGED_READ_RULES_JSON
+_ENV_ENTERPRISE_SECRET_ROTATION_DAYS = _runtime_config._ENV_ENTERPRISE_SECRET_ROTATION_DAYS
+_ENV_SWITCH_DISABLED_DEFAULT = _runtime_config._ENV_SWITCH_DISABLED_DEFAULT
+_MISSING_POLICY_VERSION_ISSUE = _runtime_config._MISSING_POLICY_VERSION_ISSUE
+_MISSING_PRIMARY_KEY_ID_ISSUE = _runtime_config._MISSING_PRIMARY_KEY_ID_ISSUE
+_RUNTIME_CONFIG_INVALID_PREFIX = _runtime_config._RUNTIME_CONFIG_INVALID_PREFIX
+_SECRET_ROTATION_DAYS_OUT_OF_RANGE_ISSUE = _runtime_config._SECRET_ROTATION_DAYS_OUT_OF_RANGE_ISSUE
+_configured_enterprise_policy_version = _runtime_config._configured_enterprise_policy_version
+_empty_json_map = _runtime_config._empty_json_map
+_enterprise_runtime_config_issues = _runtime_config._enterprise_runtime_config_issues
+_env_enabled = _runtime_config._env_enabled
+_env_int = _runtime_config._env_int
+_env_value = _runtime_config._env_value
+_load_json_map = _runtime_config._load_json_map
+_max_write_payload_bytes = _runtime_config._max_write_payload_bytes
+_normalized_enterprise_policy_version = _runtime_config._normalized_enterprise_policy_version
+_parse_int_or_default = _runtime_config._parse_int_or_default
+_primary_key_configured = _runtime_config._primary_key_configured
+_privileged_read_authz_enabled = _runtime_config._privileged_read_authz_enabled
+_runtime_config_enforcement_enabled = _runtime_config._runtime_config_enforcement_enabled
+_runtime_config_invalid_message = _runtime_config._runtime_config_invalid_message
+_write_authz_enabled = _runtime_config._write_authz_enabled
+enterprise_policy_version = _runtime_config.enterprise_policy_version
+validate_enterprise_runtime_config = _runtime_config.validate_enterprise_runtime_config
 
 logger = logging.getLogger("enterprise_readiness")
 
@@ -13,7 +52,6 @@ _SERVICE_NAME = "lotus-performance"
 _ENTERPRISE_AUDIT_EVENT_NAME = "enterprise_audit_event"
 _ENTERPRISE_AUDIT_EXTRA_KEY = "audit"
 _ENTERPRISE_POLICY_VERSION_HEADER = "X-Enterprise-Policy-Version"
-_DEFAULT_ENTERPRISE_POLICY_VERSION = "1.0.0"
 _AUDIT_PAYLOAD_SERVICE_KEY = "service"
 _AUDIT_PAYLOAD_ACTION_KEY = "action"
 _AUDIT_PAYLOAD_ACTOR_ID_KEY = "actor_id"
@@ -40,11 +78,6 @@ _REDACTED_VALUE = "***REDACTED***"
 _MISSING_HEADERS_REASON = "missing_headers"
 _MISSING_SERVICE_IDENTITY_REASON = "missing_service_identity"
 _MISSING_CAPABILITY_REASON = "missing_capability"
-_MISSING_POLICY_VERSION_ISSUE = "missing_policy_version"
-_SECRET_ROTATION_DAYS_OUT_OF_RANGE_ISSUE = "secret_rotation_days_out_of_range"
-_MISSING_PRIMARY_KEY_ID_ISSUE = "missing_primary_key_id"
-_RUNTIME_CONFIG_INVALID_PREFIX = "enterprise_runtime_config_invalid"
-_DIAGNOSTIC_LIST_SEPARATOR = ","
 _CAPABILITIES_HEADER = "x-capabilities"
 _SERVICE_IDENTITY_HEADER = "x-service-identity"
 _AUTHORIZATION_HEADER = "authorization"
@@ -58,22 +91,6 @@ _CORRELATION_ID_HEADER = "x-correlation-id"
 _UNKNOWN_ACTOR_ID = "unknown"
 _DEFAULT_TENANT_ID = "default"
 _UNKNOWN_ROLE = "unknown"
-_ENV_ENTERPRISE_ENFORCE_PRIVILEGED_READ_AUTHZ = "ENTERPRISE_ENFORCE_PRIVILEGED_READ_AUTHZ"
-_ENV_ENTERPRISE_ENFORCE_AUTHZ = "ENTERPRISE_ENFORCE_AUTHZ"
-_ENV_ENTERPRISE_ENFORCE_RUNTIME_CONFIG = "ENTERPRISE_ENFORCE_RUNTIME_CONFIG"
-_ENV_ENTERPRISE_PRIMARY_KEY_ID = "ENTERPRISE_PRIMARY_KEY_ID"
-_ENV_ENTERPRISE_MAX_WRITE_PAYLOAD_BYTES = "ENTERPRISE_MAX_WRITE_PAYLOAD_BYTES"
-_ENV_ENTERPRISE_POLICY_VERSION = "ENTERPRISE_POLICY_VERSION"
-_ENV_ENTERPRISE_SECRET_ROTATION_DAYS = "ENTERPRISE_SECRET_ROTATION_DAYS"
-_ENV_ENTERPRISE_FEATURE_FLAGS_JSON = "ENTERPRISE_FEATURE_FLAGS_JSON"
-_ENV_ENTERPRISE_CAPABILITY_RULES_JSON = "ENTERPRISE_CAPABILITY_RULES_JSON"
-_ENV_ENTERPRISE_PRIVILEGED_READ_RULES_JSON = "ENTERPRISE_PRIVILEGED_READ_RULES_JSON"
-_ENV_SWITCH_DISABLED_DEFAULT = "false"
-_ENV_ENABLED_VALUES = frozenset({"1", "true", "yes", "on"})
-_EMPTY_ENV_VALUE = ""
-_EMPTY_JSON_OBJECT = "{}"
-_DEFAULT_MAX_WRITE_PAYLOAD_BYTES = 1_048_576
-_DEFAULT_SECRET_ROTATION_DAYS = 90
 _CAPABILITY_OPERATIONS_RUNTIME_MANAGE = "operations.runtime.manage"
 _CAPABILITY_OPERATIONS_RUNTIME_READ = "operations.runtime.read"
 _HTTP_METHOD_DELETE = "DELETE"
@@ -142,96 +159,6 @@ def _is_write_method(method: str) -> bool:
 
 def _is_privileged_read_method(method: str) -> bool:
     return _normalized_http_method(method) == _HTTP_METHOD_GET
-
-
-def _env_value(name: str, default: str) -> str:
-    return os.getenv(name, default)
-
-
-def _env_enabled(name: str, default: str = "true") -> bool:
-    return _env_value(name, default).strip().lower() in _ENV_ENABLED_VALUES
-
-
-def _privileged_read_authz_enabled() -> bool:
-    return _env_enabled(_ENV_ENTERPRISE_ENFORCE_PRIVILEGED_READ_AUTHZ, _ENV_SWITCH_DISABLED_DEFAULT)
-
-
-def _write_authz_enabled() -> bool:
-    return _env_enabled(_ENV_ENTERPRISE_ENFORCE_AUTHZ, _ENV_SWITCH_DISABLED_DEFAULT)
-
-
-def _runtime_config_enforcement_enabled() -> bool:
-    return _env_enabled(_ENV_ENTERPRISE_ENFORCE_RUNTIME_CONFIG, _ENV_SWITCH_DISABLED_DEFAULT)
-
-
-def _primary_key_configured() -> bool:
-    return bool(_env_value(_ENV_ENTERPRISE_PRIMARY_KEY_ID, _EMPTY_ENV_VALUE).strip())
-
-
-def _max_write_payload_bytes() -> int:
-    return _env_int(_ENV_ENTERPRISE_MAX_WRITE_PAYLOAD_BYTES, _DEFAULT_MAX_WRITE_PAYLOAD_BYTES)
-
-
-def _load_json_map(name: str) -> dict[str, Any]:
-    raw = _env_value(name, _EMPTY_JSON_OBJECT)
-    try:
-        parsed = json.loads(raw)
-    except json.JSONDecodeError:
-        return _empty_json_map()
-    return parsed if isinstance(parsed, dict) else _empty_json_map()
-
-
-def _empty_json_map() -> dict[str, Any]:
-    return {}
-
-
-def _env_int(name: str, default: int) -> int:
-    return _parse_int_or_default(_env_value(name, str(default)), default)
-
-
-def _parse_int_or_default(value: Any, default: int) -> int:
-    try:
-        return int(value)
-    except (TypeError, ValueError):
-        return default
-
-
-def _configured_enterprise_policy_version() -> str:
-    return _env_value(_ENV_ENTERPRISE_POLICY_VERSION, _DEFAULT_ENTERPRISE_POLICY_VERSION)
-
-
-def _normalized_enterprise_policy_version() -> str:
-    return _configured_enterprise_policy_version().strip()
-
-
-def enterprise_policy_version() -> str:
-    return _normalized_enterprise_policy_version() or _DEFAULT_ENTERPRISE_POLICY_VERSION
-
-
-def _enterprise_runtime_config_issues() -> list[str]:
-    issues: list[str] = []
-    if not _normalized_enterprise_policy_version():
-        issues.append(_MISSING_POLICY_VERSION_ISSUE)
-
-    rotation_days = _env_int(_ENV_ENTERPRISE_SECRET_ROTATION_DAYS, _DEFAULT_SECRET_ROTATION_DAYS)
-    if rotation_days <= 0 or rotation_days > _DEFAULT_SECRET_ROTATION_DAYS:
-        issues.append(_SECRET_ROTATION_DAYS_OUT_OF_RANGE_ISSUE)
-
-    if _write_authz_enabled() and not _primary_key_configured():
-        issues.append(_MISSING_PRIMARY_KEY_ID_ISSUE)
-
-    return issues
-
-
-def _runtime_config_invalid_message(issues: list[str]) -> str:
-    return f"{_RUNTIME_CONFIG_INVALID_PREFIX}:{_DIAGNOSTIC_LIST_SEPARATOR.join(issues)}"
-
-
-def validate_enterprise_runtime_config() -> list[str]:
-    issues = _enterprise_runtime_config_issues()
-    if issues and _runtime_config_enforcement_enabled():
-        raise RuntimeError(_runtime_config_invalid_message(issues))
-    return issues
 
 
 def load_feature_flags() -> dict[str, dict[str, dict[str, bool]]]:
